+++ src/main/kotlin/com/brex/plugins/codeowners/CodeOwnersWidget.kt
@@ -96,5 +96,4 @@ class CodeOwnersWidget(project: Project) : EditorBasedWidget(project), StatusBar
             }
         }
     }
-
 }